Added amenities feature

+++ src/beans/Amenity.java
@@ -2,7 +2,7 @@ package beans;
 
 public class Amenity {
 
-	private Long id;
+	private Integer id;
 	
 	private String name;
 	
@@ -10,17 +10,17 @@ public class Amenity {
 		
 	}
 
-	public Amenity(Long id, String name) {
+	public Amenity(Integer id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
